Allowing users add/remove movies to Favorites

+++ app/src/main/java/com/pedrolopesme/android/cinepedia/dao/ReviewDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ReviewDao {
 
-    List<Review> get(int movieId);
+    List<Review> get(long movieId);
 
 }
